Cleanup

+++ 2D_Game_Engine/Base_Classes/timer.cpp
@@ -1,7 +1,7 @@
 #include "timer.h"
 #include <SDL2/SDL.h>
 #include<iostream>
-#include "coreengine.h"
+#include "window.h"
 
 Timer::Timer()
 {
@@ -9,14 +9,11 @@ Timer::Timer()
     totalTime=0;
 }
 
-int Timer::printFPS()
+double Timer::printFPS()
 {
-    const Uint32 startTime = SDL_GetTicks();
+    const double startTime = SDL_GetTicks();
     SDL_Delay( 1.0f/60.0f );
-    const Uint32 endTime = SDL_GetTicks();
-
+    const double endTime = SDL_GetTicks();
     totalTime = endTime - startTime;
-    //++numFrames;
-    //float fps = numFrames / (totalTime / 1000.0);
     return totalTime;
 }
